backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI/ML Libraries
google-generativeai==0.3.1
huggingface-hub==0.19.4
transformers==4.36.0
torch==2.1.2

# Caching & Database
redis==5.0.1
diskcache==5.6.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
tenacity==8.2.3

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
